+++ src/main/java/com/cake/bnd/foundation/config/conditions/BndNotSuppressingCondition.java
@@ -1,0 +1,29 @@
+package com.cake.bnd.foundation.config.conditions;
+
+import com.cake.bnd.foundation.BndSuppression;
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.MapCodec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.neoforged.neoforge.common.conditions.ICondition;
+import org.jetbrains.annotations.NotNull;
+
+public record BndNotSuppressingCondition(String category) implements ICondition {
+
+    public static final MapCodec<BndNotSuppressingCondition> CODEC = RecordCodecBuilder.mapCodec(
+            c ->
+                    c.group(
+                            Codec.STRING.fieldOf("category").forGetter(BndNotSuppressingCondition::category)
+                    ).apply(c, BndNotSuppressingCondition::new)
+    );
+
+    @Override
+    public boolean test(final @NotNull IContext context) {
+        return BndSuppression.shouldLoadRecipes(this.category);
+    }
+
+    @Override
+    public @NotNull MapCodec<? extends ICondition> codec() {
+        return CODEC;
+    }
+
+}