+++ src/main/java/com/cake/bnd/registry/client/BndPartialModels.java
@@ -1,0 +1,13 @@
+package com.cake.bnd.registry.client;
+
+import com.simibubi.create.Create;
+import dev.engine_room.flywheel.lib.model.baked.PartialModel;
+
+public class BndPartialModels {
+
+    public static final PartialModel FROGPORT_ITEM = PartialModel.of(Create.asResource("block/package_frogport/item"));
+
+    public static void register() {
+    }
+
+}