+++ changelog.md
@@ -17,3 +17,9 @@
 # 1.21.1-1.0.4
 
 - Added vault dye capability (Including compat for Create: Connected)
+
+# 1.21.1-1.1.0
+
+- Added vivid vaults texture pack for more saturated vault palette colors
+- Added compat with Create: Vibrant Vaults
+- Added dyeable functionality for Frogports and Redstone Requester